Implement action button timer logic
I implemented this a while ago, but it wasn't working and I couldn't,
for the life of me, understand why. Now, with the simulation, I *know*
that it's not some glaring, stupid logic error.

It has to be a hardware problem or a bad register manipulation in the
code abstracted away by the sim.