FINALLY fixed that goddamn bug!
The root cause was that I forgot to set my interrupt on timer 1 when
I changed my timer for 0 to 1 a while ago. I didn't notice it because
you know what? The program still worked. You know why? Because whenever
the timer matched and try to trigger the interrupt, **it would reset the
MCU**.

Nuggets of wisdom for future debugging: Program is behaving weirdly? Two
things to check first:

1. Did you smash your stack?
2. Is your MCU resetting at some point without you noticing?